+++ docs/quick_start/quick_start_with_jdbc.md
@@ -68,7 +68,7 @@ import org.apache.hadoop.security.UserGroupInformation;
  
 public class JDBCTest {
  
-    private static String driverName = "org.apache.hive.jdbc.KyuubiHiveDriver";
+    private static String driverName = "KyuubiHiveDriver";
     private static String kyuubiJdbcUrl = "jdbc:hive2://localhost:10009/default;";
  
     public static void main(String[] args) throws ClassNotFoundException, SQLException {
