+++ kyuubi-hive-jdbc/src/main/java/org/apache/kyuubi/jdbc/hive/ZooKeeperHiveClientHelper.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package org.apache.hive.jdbc;
+package org.apache.kyuubi.jdbc.hive;
 
 import java.nio.charset.Charset;
 import java.util.List;
@@ -27,7 +27,6 @@ import java.util.regex.Pattern;
 import org.apache.curator.framework.CuratorFramework;
 import org.apache.curator.framework.CuratorFrameworkFactory;
 import org.apache.curator.retry.ExponentialBackoffRetry;
-import org.apache.hive.jdbc.Utils.JdbcConnectionParams;
 import org.apache.zookeeper.Watcher;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -45,13 +44,13 @@ class ZooKeeperHiveClientHelper {
     }
   }
 
-  static void configureConnParams(JdbcConnectionParams connParams)
+  static void configureConnParams(Utils.JdbcConnectionParams connParams)
       throws ZooKeeperHiveClientException {
     String zooKeeperEnsemble = connParams.getZooKeeperEnsemble();
     String zooKeeperNamespace =
-        connParams.getSessionVars().get(JdbcConnectionParams.ZOOKEEPER_NAMESPACE);
+        connParams.getSessionVars().get(Utils.JdbcConnectionParams.ZOOKEEPER_NAMESPACE);
     if ((zooKeeperNamespace == null) || (zooKeeperNamespace.isEmpty())) {
-      zooKeeperNamespace = JdbcConnectionParams.ZOOKEEPER_DEFAULT_NAMESPACE;
+      zooKeeperNamespace = Utils.JdbcConnectionParams.ZOOKEEPER_DEFAULT_NAMESPACE;
     }
     List<String> serverHosts;
     Random randomizer = new Random();
@@ -110,7 +109,7 @@ class ZooKeeperHiveClientHelper {
    * @param connParams
    * @throws Exception
    */
-  private static void applyConfs(String serverConfStr, JdbcConnectionParams connParams)
+  private static void applyConfs(String serverConfStr, Utils.JdbcConnectionParams connParams)
       throws Exception {
     Matcher matcher = kvPattern.matcher(serverConfStr);
     while (matcher.find()) {
@@ -126,8 +125,8 @@ class ZooKeeperHiveClientHelper {
         }
         // Set transportMode
         if ((matcher.group(1).equals("hive.server2.transport.mode"))
-            && !(connParams.getSessionVars().containsKey(JdbcConnectionParams.TRANSPORT_MODE))) {
-          connParams.getSessionVars().put(JdbcConnectionParams.TRANSPORT_MODE, matcher.group(2));
+            && !(connParams.getSessionVars().containsKey(Utils.JdbcConnectionParams.TRANSPORT_MODE))) {
+          connParams.getSessionVars().put(Utils.JdbcConnectionParams.TRANSPORT_MODE, matcher.group(2));
         }
         // Set port
         if (matcher.group(1).equals("hive.server2.thrift.port")) {
@@ -139,18 +138,18 @@ class ZooKeeperHiveClientHelper {
         }
         // Set sasl qop
         if ((matcher.group(1).equals("hive.server2.thrift.sasl.qop"))
-            && !(connParams.getSessionVars().containsKey(JdbcConnectionParams.AUTH_QOP))) {
-          connParams.getSessionVars().put(JdbcConnectionParams.AUTH_QOP, matcher.group(2));
+            && !(connParams.getSessionVars().containsKey(Utils.JdbcConnectionParams.AUTH_QOP))) {
+          connParams.getSessionVars().put(Utils.JdbcConnectionParams.AUTH_QOP, matcher.group(2));
         }
         // Set http path
         if ((matcher.group(1).equals("hive.server2.thrift.http.path"))
-            && !(connParams.getSessionVars().containsKey(JdbcConnectionParams.HTTP_PATH))) {
-          connParams.getSessionVars().put(JdbcConnectionParams.HTTP_PATH, matcher.group(2));
+            && !(connParams.getSessionVars().containsKey(Utils.JdbcConnectionParams.HTTP_PATH))) {
+          connParams.getSessionVars().put(Utils.JdbcConnectionParams.HTTP_PATH, matcher.group(2));
         }
         // Set SSL
         if ((matcher.group(1) != null) && (matcher.group(1).equals("hive.server2.use.SSL"))
-            && !(connParams.getSessionVars().containsKey(JdbcConnectionParams.USE_SSL))) {
-          connParams.getSessionVars().put(JdbcConnectionParams.USE_SSL, matcher.group(2));
+            && !(connParams.getSessionVars().containsKey(Utils.JdbcConnectionParams.USE_SSL))) {
+          connParams.getSessionVars().put(Utils.JdbcConnectionParams.USE_SSL, matcher.group(2));
         }
         /**
          * Note: this is pretty messy, but sticking to the current implementation.
@@ -165,21 +164,21 @@ class ZooKeeperHiveClientHelper {
         if (matcher.group(1).equals("hive.server2.authentication")) {
           // NOSASL
           if (matcher.group(2).equalsIgnoreCase("NOSASL")
-              && !(connParams.getSessionVars().containsKey(JdbcConnectionParams.AUTH_TYPE) && connParams
-                  .getSessionVars().get(JdbcConnectionParams.AUTH_TYPE)
-                  .equalsIgnoreCase(JdbcConnectionParams.AUTH_SIMPLE))) {
-            connParams.getSessionVars().put(JdbcConnectionParams.AUTH_TYPE,
-                JdbcConnectionParams.AUTH_SIMPLE);
+              && !(connParams.getSessionVars().containsKey(Utils.JdbcConnectionParams.AUTH_TYPE) && connParams
+                  .getSessionVars().get(Utils.JdbcConnectionParams.AUTH_TYPE)
+                  .equalsIgnoreCase(Utils.JdbcConnectionParams.AUTH_SIMPLE))) {
+            connParams.getSessionVars().put(Utils.JdbcConnectionParams.AUTH_TYPE,
+                Utils.JdbcConnectionParams.AUTH_SIMPLE);
           }
         }
         // KERBEROS
         // If delegation token is passed from the client side, do not set the principal
         if (matcher.group(1).equalsIgnoreCase("hive.server2.authentication.kerberos.principal")
-            && !(connParams.getSessionVars().containsKey(JdbcConnectionParams.AUTH_TYPE) && connParams
-                .getSessionVars().get(JdbcConnectionParams.AUTH_TYPE)
-                .equalsIgnoreCase(JdbcConnectionParams.AUTH_TOKEN))
-            && !(connParams.getSessionVars().containsKey(JdbcConnectionParams.AUTH_PRINCIPAL))) {
-          connParams.getSessionVars().put(JdbcConnectionParams.AUTH_PRINCIPAL, matcher.group(2));
+            && !(connParams.getSessionVars().containsKey(Utils.JdbcConnectionParams.AUTH_TYPE) && connParams
+                .getSessionVars().get(Utils.JdbcConnectionParams.AUTH_TYPE)
+                .equalsIgnoreCase(Utils.JdbcConnectionParams.AUTH_TOKEN))
+            && !(connParams.getSessionVars().containsKey(Utils.JdbcConnectionParams.AUTH_PRINCIPAL))) {
+          connParams.getSessionVars().put(Utils.JdbcConnectionParams.AUTH_PRINCIPAL, matcher.group(2));
         }
       }
     }
